Proper address module
